refactor: move the database client creation into the common module
