+++ modules/common/package.json
@@ -33,6 +33,7 @@
   },
   "dependencies": {
     "@mikro-orm/core": "^5.7.12",
+    "@mikro-orm/reflection": "^5.7.12",
     "ajv": "^8.12.0",
     "ajv-errors": "^3.0.0",
     "node-fetch": "^3.3.0"
